requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt --python-platform=linux
agate==1.9.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
annotated-types==0.7.0
    # via pydantic
appdirs==1.4.4
    # via sqlfluff
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
babel==2.16.0
    # via agate
cachetools==5.5.0
    # via google-auth
certifi==2024.7.4
    # via requests
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via
    #   diff-cover
    #   sqlfluff
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
    #   sqlfluff
    #   typer
colorama==0.4.6
    # via
    #   dbt-common
    #   sqlfluff
daff==1.3.46
    # via dbt-core
db-dtypes==1.3.0
    # via google-cloud-bigquery
dbt-adapters==1.4.1
    # via
    #   dbt-bigquery
    #   dbt-core
dbt-bigquery==1.8.1
    # via -r requirements.in
dbt-common==1.7.0
    # via
    #   dbt-adapters
    #   dbt-bigquery
    #   dbt-core
dbt-core==1.8.6
    # via
    #   -r requirements.in
    #   dbt-bigquery
    #   sqlfluff-templater-dbt
dbt-extractor==0.5.1
    # via dbt-core
dbt-semantic-interfaces==0.5.1
    # via dbt-core
deepdiff==7.0.1
    # via dbt-common
diff-cover==9.1.1
    # via sqlfluff
distlib==0.3.8
    # via virtualenv
exceptiongroup==1.2.2
    # via pytest
filelock==3.15.4
    # via virtualenv
future==1.0.0
    # via parsedatetime
google-api-core==2.19.1
    # via
    #   dbt-bigquery
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-dataproc
    #   google-cloud-storage
google-auth==2.34.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-dataproc
    #   google-cloud-storage
google-cloud-bigquery==3.25.0
    # via dbt-bigquery
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-dataproc==5.10.2
    # via dbt-bigquery
google-cloud-storage==2.18.2
    # via dbt-bigquery
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.63.2
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-dataproc
grpcio==1.66.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
identify==2.6.0
    # via pre-commit
idna==3.8
    # via requests
importlib-metadata==6.11.0
    # via dbt-semantic-interfaces
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via
    #   agate
    #   dbt-common
jinja2==3.1.4
    # via
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   diff-cover
    #   jinja2-simple-tags
    #   sqlfluff
jinja2-simple-tags==0.6.1
    # via sqlfluff-templater-dbt
jsonschema==4.23.0
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
jsonschema-specifications==2023.12.1
    # via jsonschema
leather==0.4.0
    # via agate
logbook==1.5.3
    # via dbt-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mashumaro==3.13.1
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
mdurl==0.1.2
    # via markdown-it-py
minimal-snowplow-tracker==0.0.2
    # via dbt-core
more-itertools==10.5.0
    # via dbt-semantic-interfaces
msgpack==1.0.8
    # via mashumaro
networkx==2.8.8
    # via dbt-core
nodeenv==1.9.1
    # via pre-commit
numpy==2.1.1
    # via
    #   db-dtypes
    #   pandas
    #   pyarrow
ordered-set==4.1.0
    # via deepdiff
packaging==24.1
    # via
    #   db-dtypes
    #   dbt-core
    #   google-cloud-bigquery
    #   pytest
pandas==2.2.2
    # via
    #   db-dtypes
    #   google-cloud-bigquery
parsedatetime==2.4
    # via agate
pathspec==0.11.2
    # via
    #   dbt-common
    #   dbt-core
    #   sqlfluff
platformdirs==4.2.2
    # via virtualenv
pluggy==1.5.0
    # via
    #   diff-cover
    #   pytest
pre-commit==2.21.0
    # via -r requirements.in
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-dataproc
protobuf==4.25.4
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   google-api-core
    #   google-cloud-dataproc
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyarrow==17.0.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.9.1
    # via dbt-semantic-interfaces
pydantic-core==2.23.3
    # via pydantic
pygments==2.18.0
    # via
    #   diff-cover
    #   rich
pytest==8.3.2
    # via sqlfluff
python-dateutil==2.8.2
    # via
    #   dbt-common
    #   dbt-semantic-interfaces
    #   google-cloud-bigquery
    #   pandas
python-slugify==8.0.4
    # via agate
pytimeparse==1.1.8
    # via agate
pytz==2024.1
    # via
    #   dbt-adapters
    #   dbt-core
    #   pandas
pyyaml==6.0.2
    # via
    #   dbt-core
    #   dbt-semantic-interfaces
    #   pre-commit
    #   sqlfluff
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via sqlfluff
requests==2.32.3
    # via
    #   dbt-common
    #   dbt-core
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   minimal-snowplow-tracker
rich==13.7.1
    # via typer
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
ruff==0.5.7
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   isodate
    #   minimal-snowplow-tracker
    #   python-dateutil
sqlfluff==3.1.1
    # via
    #   -r requirements.in
    #   sqlfluff-templater-dbt
sqlfluff-templater-dbt==3.1.1
    # via -r requirements.in
sqlparse==0.5.1
    # via dbt-core
tblib==3.0.0
    # via sqlfluff
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via sqlfluff
tomli==2.0.1
    # via pytest
tqdm==4.66.5
    # via sqlfluff
typer==0.12.3
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   dbt-adapters
    #   dbt-common
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via requests
virtualenv==20.26.3
    # via pre-commit
zipp==3.20.1
    # via importlib-metadata
